Move skill content loading into launch resolve
